Add a converter class for generating strings for call parameter sets.
